deletes unnecessary files and adds integrationtests package comment
