+++ .gitignore
@@ -26,8 +26,7 @@ website/node_modules
 *.test
 *.iml
 *.lock.hcl
-*state.json
-terraform-provider-grid/pkg/*
+state.json
 
 website/vendor
